refactor(abw): split youtube SEO stage concerns
