+++ apps/ai-blog-writer/apps/backend/app/features/youtube2blog/stages/stage_seo_prompts.py
@@ -1,0 +1,78 @@
+"""Prompt policy for YouTube2Blog SEO brief and enrichment phases."""
+
+SEO_BRIEF_PROMPT = """You are an SEO strategist for editorial content.
+
+Task:
+- Infer on-page SEO direction from the article draft and article type.
+- Choose keywords automatically (no user-supplied keyword list).
+- Keep choices realistic and semantically aligned with the draft.
+
+ARTICLE TITLE:
+{title}
+
+ARTICLE TYPE:
+{article_type}
+
+GUIDELINE:
+{guideline}
+
+ARTICLE DRAFT (MARKDOWN):
+{article}
+
+Return strict JSON only:
+{{
+  "search_intent": "<informational|commercial investigation|navigational|transactional>",
+  "focus_keyword": "<1 short phrase>",
+  "secondary_keywords": ["<5 to 8 related phrases>"],
+  "seo_objective": "<1 concise sentence>",
+  "heading_hints": ["<3 to 5 SEO-friendly heading ideas>"]
+}}
+
+Constraints:
+- Focus keyword must be naturally supported by the draft.
+- Secondary keywords must be semantically related and non-duplicative.
+- Avoid vague generic keywords.
+"""
+
+SEO_ENRICH_PROMPT = """You are improving on-page SEO quality of an article draft.
+
+Mode: {mode}
+Search intent: {search_intent}
+Focus keyword: {focus_keyword}
+Focus keyword count in source: {focus_count_before}
+Max focus keyword count in output: {max_focus_occurrences}
+Max focus keyword increase vs source: {max_focus_increase}
+Max focus keyword density: {max_focus_density_pct}%
+Secondary keywords: {secondary_keywords}
+SEO objective: {seo_objective}
+Heading hints: {heading_hints}
+
+Article type: {article_type}
+Guideline: {guideline}
+
+Write for readers first and SEO second. Use natural travel-news language, avoid keyword stuffing, avoid repetitive SEO headings, and make the article feel edited by a human. Include SEO elements only where they improve clarity: a strong headline, concise subhead, clean section structure, accurate metadata, and natural keywords. SEO structure and keywords never override the voice rules appended below.
+
+Rules:
+1. Preserve factual meaning from the source draft.
+2. Do NOT invent facts, numbers, names, dates, or claims.
+3. Keep the writing natural and useful (human-first). SEO is secondary to quality.
+4. Use the focus keyword naturally, without forcing repetition.
+5. Prefer subtle optimization over aggressive rewriting.
+6. Do not overuse secondary keywords; include only when they fit naturally.
+7. Preserve markdown structure and readability.
+8. Keep overall coverage depth; do not collapse the article.
+9. Hard safety limits:
+   - Never exceed the max focus keyword count.
+   - Never exceed the max focus keyword increase.
+   - Keep focus keyword density below the max density.
+
+Retry feedback (if present):
+{feedback}
+
+SOURCE ARTICLE:
+{article}
+
+Output only improved markdown. No JSON. No explanations.
+"""
+
+__all__ = ["SEO_BRIEF_PROMPT", "SEO_ENRICH_PROMPT"]
